requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
numpy
pypdf
python-docx
python-dotenv